stop being stupid vscode please
